delete reasoning-memory.test.ts and move test into agent-memory.test.ts instead
